python 3 compatibility
This change allows PyMKS to be compatible on Python 3.5.

It changes the way PyMKS uses print statements to work with Python 3 syntax, `range()`, and casts some function input arguments to ints.

All tests ran successfully on my machine minus the expected sfepy failures. All examples on the PyMKS website that do not use sfepy also run successfully.

+++ pymks/datasets/__init__.py
@@ -252,7 +252,7 @@ def make_checkerboard_microstructure(square_size, n_squares):
     X = np.ones((2 * square_size, 2 * square_size), dtype=int)
     X[:square_size, :square_size] = 0
     X[square_size:, square_size:] = 0
-    return np.tile(X, ((n_squares + 1) / 2, (n_squares + 1) / 2))[None, :L, :L]
+    return np.tile(X, (int((n_squares + 1) / 2), int((n_squares + 1) / 2)))[None, :L, :L]
 
 
 def make_elastic_stress_random(n_samples=[10, 10], elastic_modulus=(100, 150),
